[gfx] update Truffel design
Truffel, the cat of Saffron appears on a single panel in the 'Potion Tournament'. The initial character design was a machiavellian, interested and tricky cat , a sort of total opposite of Carrot. Almost white with a brown stain on one eyes, and thin eye. But I decided animals in Pepper&Carrot shouldn't be evil. Especially companion creatures.  Coriander as a grumpy rooster, Shichimi has a goofy fox, Pepper has the innocent Carrot ... I was needing a female : Truffel for Saffron.
Fixing Truffel was not a big deal and doesn't affect the story at all : removing the stain around eyes, and add eye-lash. I do it for future consistency. 

( note : I think [gfx] modification is something that should be kept as rare and exeptionnal as possible )